+++ packages/web-components/src/components/global-header/components/global-header/src/globals/loadSolisScript.ts
@@ -41,8 +41,11 @@ export default function loadSolisScript(props: HeaderProps) {
         cdn_hostname: props.solisConfig.cdn_hostname,
         deployment_environment: props.solisConfig.deployment_environment,
         backend_proxy: props.solisConfig.backendProxy,
+        coachmark_enabled: false,
       };
     }
+
+    window._solis.custom_switcher_button = true;
 
     const setAttributeFromEvent = (event: { type: string }) => {
       script?.setAttribute(