+++ packages/web-components/src/components/global-header/components/global-header/src/types/Header.types.ts
@@ -131,6 +131,7 @@ export type ProfileFooterLinks = {
   carbonIcon?: string;
   newTab?: boolean;
   newTabIcon?: boolean;
+  'data-ypc-link'?: boolean;
   onClickHandler?: () => void;
 };
 
@@ -314,6 +315,7 @@ export interface ReactWrapperProps extends Omit<
   productVersion?: null;
   assistMeKey?: string;
   hasNewNotifications?: boolean;
+  forceBackendProxy?: boolean;
   capabilityProfileFooterLinks?: ProfileFooterLinks[];
   capabilityGlobalActions?: GlobalActionConfig[];
   searchConfigs?: Omit<SearchConfigs, 'callback' | 'submitCallback'>;
@@ -362,6 +364,13 @@ export type SidekickInfo = {
   chat_enabled?: boolean;
   overview_enabled?: boolean;
   tell_me_more_enabled?: boolean;
+  openSidekick?: () => void | undefined;
+  closeSidekick?: () => void | undefined;
+};
+
+export type SwitcherInfo = {
+  openSwitcher?: () => void | undefined;
+  closeSwitcher?: () => void | undefined;
 };
 
 interface solisWindowConfig {
@@ -369,12 +378,22 @@ interface solisWindowConfig {
   cdn_hostname: string;
   deployment_environment: solisDeploymentEnvironment;
   sidekick?: SidekickInfo;
+  switcher?: SwitcherInfo;
   product_id: string;
   backend_proxy?: string;
+  coachmark_enabled?: boolean;
+  custom_switcher_button?: boolean;
+  custom_sidekick_button?: boolean;
 }
 
 declare global {
   interface Window {
     _solis: solisWindowConfig;
   }
+}
+
+export interface solisSessionManagerConfig {
+  tokenRefreshInterval?: number;
+  idleTimeoutInterval?: number;
+  basePath?: string;
 }