+++ packages/web-components/src/components/chat/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.39.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/ai-chat@0.38.0...@carbon-labs/ai-chat@0.39.0) (2026-06-02)
+
+**Note:** Version bump only for package @carbon-labs/ai-chat
+
+
+
+
+
 # [0.38.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/ai-chat@0.37.0...@carbon-labs/ai-chat@0.38.0) (2026-05-20)
 
 **Note:** Version bump only for package @carbon-labs/ai-chat